+++ Tests/Editor/Scripts/SceneComponentsTest.cs
@@ -11,15 +11,17 @@ internal class SceneComponentsTest {
     [UnityTest]
     public IEnumerator EnsureClearOnSceneClosed() {
 
-        DestroyAllLights(); //Clear all lights first
-                
         const int NUM_LIGHTS = 3;
+        
+        DestroyAllLights(); //Clear all lights first
+
+        SceneComponents<Light> lightComponents = SceneComponents<Light>.GetInstance();
         for (int i = 0; i < NUM_LIGHTS; ++i) {
             new GameObject().AddComponent<Light>();
-            SceneComponents<Light>.GetInstance().Update();
+            lightComponents.Update();
             yield return null;
             
-            Assert.AreEqual(i + 1 , SceneComponents<Light>.GetInstance().GetCachedComponents().Count);
+            Assert.AreEqual(i + 1 , lightComponents.GetCachedComponents().Count);
         }
 
         yield return null;
@@ -27,7 +29,7 @@ internal class SceneComponentsTest {
         EditorSceneManager.NewScene(NewSceneSetup.EmptyScene, NewSceneMode.Single);
         yield return null;
         
-        Assert.AreEqual(0 , SceneComponents<Light>.GetInstance().GetCachedComponents().Count);
+        Assert.AreEqual(0 , lightComponents.GetCachedComponents().Count);
 
     }
     
@@ -35,29 +37,64 @@ internal class SceneComponentsTest {
     
     [Test]
     public void CompareUpdateMethods() {
+        const int NUM_LIGHTS = 3;
         DestroyAllLights(); //Clear all lights first
         
-        SceneComponents<Light>.GetInstance().Update();
-        Assert.AreEqual(0 , SceneComponents<Light>.GetInstance().GetCachedComponents().Count);
+        SceneComponents<Light> lightComponents = SceneComponents<Light>.GetInstance();
+        lightComponents.Update();
+        Assert.AreEqual(0 , lightComponents.GetCachedComponents().Count);
         
-        const int NUM_LIGHTS = 3;
         for (int i = 0; i < NUM_LIGHTS; ++i) {
             new GameObject().AddComponent<Light>();
-            SceneComponents<Light>.GetInstance().Update();
-            Assert.AreEqual(i , SceneComponents<Light>.GetInstance().GetCachedComponents().Count);
+            lightComponents.Update();
+            Assert.AreEqual(i , lightComponents.GetCachedComponents().Count);
             
-            SceneComponents<Light>.GetInstance().ForceUpdate();
-            Assert.AreEqual(i + 1 , SceneComponents<Light>.GetInstance().GetCachedComponents().Count);
+            lightComponents.ForceUpdate();
+            Assert.AreEqual(i + 1 , lightComponents.GetCachedComponents().Count);
         }
     }
 
-//--------------------------------------------------------------------------------------------------------------------------------------------------------------    
+//--------------------------------------------------------------------------------------------------------------------------------------------------------------
     
-    void DestroyAllLights() {
-        foreach (Light l in Object.FindObjectsByType<Light>(FindObjectsSortMode.None)) {
-            Object.DestroyImmediate(l.gameObject);
-        }
+    [Test]
+    public void IncludeDisabledObjects() {
+        const int NUM_LIGHTS = 3;
+        DestroyAllLights(); //Clear all lights first
         
+        SceneComponents<Light> lightComponents = SceneComponents<Light>.GetInstance();
+        lightComponents.SetIncludeInactive(true);
+        CreateLightObjects(NUM_LIGHTS, active:false);
+        
+        lightComponents.ForceUpdate();
+        Assert.AreEqual(NUM_LIGHTS , lightComponents.GetCachedComponents().Count);
+    }
+
+    [Test]
+    public void ExcludeDisabledObjects() {
+        DestroyAllLights(); //Clear all lights first
+        
+        SceneComponents<Light> lightComponents = SceneComponents<Light>.GetInstance();
+        lightComponents.SetIncludeInactive(false);
+        CreateLightObjects(numLights:3, active:false);
+        
+        lightComponents.ForceUpdate();
+        Assert.AreEqual(0 , lightComponents.GetCachedComponents().Count);
+    }
+
+//--------------------------------------------------------------------------------------------------------------------------------------------------------------
+
+    static void CreateLightObjects(int numLights, bool active) {
+        for (int i = 0; i < numLights; ++i) {
+            GameObject go = new GameObject();
+            go.AddComponent<Light>();
+            go.SetActive(active);
+        }
+    }
+    
+    static void DestroyAllLights() {
+        Object.FindObjectsByType<Light>(FindObjectsInactive.Include, FindObjectsSortMode.None).Loop((Light l) => {
+            Object.DestroyImmediate(l.gameObject);
+        });
     }
     
 }